+++ frontend/src/css/index.css
@@ -1,46 +1,147 @@
 body {
-    margin: 0%;
-    padding: 0%;
+  margin: 0;
+  padding: 0;
 }
 
 .App {
-    display: flex;
+  display: flex;
+  flex-direction: column;
+  background-color: black;
+  padding-bottom: 7.84vw;
+}
+.App .EventCards {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: 10px;
+  margin: auto;
+}
+@media (max-width: 1023px) {
+  .App .EventCards {
     flex-direction: column;
-    background-color: black;
-    padding-bottom: 7.84vw;
-}
-
-.EventCards {
-    display: block;
-    width: auto;
-    list-style-type: disc;
-    margin-block-start: 1em;
-    margin-block-end: 1em;
-    margin-inline-start: 0px;
-    margin-inline-end: 0px;
-    margin: auto;
-    padding-inline-start: 40px;
-}
-
-*,
-:after,
-:before {
-    box-sizing: border-box;
+  }
 }
 
 .Events__heading {
-    font-size: xxx-large;
-    text-transform: uppercase;
-    background: linear-gradient(#f9daa8, #bc615e);
-    color: transparent;
-    -webkit-background-clip: text;
-    background-clip: text;
-    margin: auto;
-    font-family: "Carta Marina W00 Regular";
+  font-size: xxx-large;
+  text-transform: uppercase;
+  background: linear-gradient(#f9daa8, #bc615e);
+  color: transparent;
+  -webkit-background-clip: text;
+  background-clip: text;
+  margin: auto;
+  font-family: "Carta Marina W00 Regular";
 }
 
-@media (min-width: 768px) {
-    .EventCards {
-        max-width: 83.61vw;
-    }
+.EventCard {
+  display: flex;
+  flex-direction: column;
+  width: auto;
+  margin-bottom: 5.73vw;
+}
+.EventCard.hiddenCard {
+  opacity: 0;
+  transform: translateY(90px) scale(0.9);
+  transition: all 0.5s;
+}
+.EventCard.showCard {
+  opacity: 1;
+  transform: translateY(0px) scale(1);
+}
+.EventCard .event-container {
+  height: 100%;
+  padding: 1.5rem;
+  width: fit-content;
+  margin: 0 auto;
+}
+.EventCard .event-container .event__img-container {
+  overflow: hidden;
+  width: 35.16vw;
+  height: 38.49vw;
+}
+@media (max-width: 1023px) {
+  .EventCard .event-container .event__img-container {
+    width: 78.54vw;
+    height: 85.97vw;
+  }
+}
+.EventCard .event-container .event__img-container .event__img {
+  margin: auto;
+  width: 100%;
+  height: 100%;
+  transition: all 0.8s;
+}
+.EventCard .event-container .event__img-container .event__img:hover {
+  transform: scale(1.1);
+}
+.EventCard .event-container .event__content {
+  display: flex;
+  font-family: sans-serif;
+  flex-direction: column;
+  align-items: flex-start;
+  color: #fff;
+  width: 35.16vw;
+}
+.EventCard .event-container .event__content .event__name {
+  margin-bottom: 0;
+  font-size: 2.5vw;
+  line-height: 1.51vw;
+  margin-top: 1.93vw;
+  font-weight: 50;
+  text-transform: uppercase;
+}
+@media (max-width: 1023px) {
+  .EventCard .event-container .event__content .event__name {
+    font-size: 5.18vw;
+    line-height: 3.37vw;
+    margin-top: 4.31vw;
+  }
+}
+.EventCard .event-container .event__content .event__desc {
+  margin: 0;
+  font-size: 1.5vw;
+  color: rgb(144, 144, 144);
+  line-height: 0.99vw;
+  margin-top: 0.57vw;
+  font-weight: 100;
+}
+@media (max-width: 1023px) {
+  .EventCard .event-container .event__content .event__desc {
+    font-size: 3.09vw;
+    line-height: 2.23vw;
+    margin-top: 1.77vw;
+  }
+}
+@media (min-width: 1023px) {
+  .EventCard:nth-child(2n) {
+    margin-left: 4.43vw;
+  }
+  .EventCard:nth-child(2n).hiddenCard {
+    transform: translateY(calc(90px + 9.11vw)) scale(0.9);
+  }
+  .EventCard:nth-child(2n).showCard {
+    transform: translateY(9.11vw) scale(1);
+  }
+}
+
+.explore {
+  color: #fff;
+  width: 35.16vw;
+  height: 38.49vw;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-family: sans-serif;
+  font-size: 2.5vw;
+}
+.explore__text {
+  width: fit-content;
+}
+@media (max-width: 1023px) {
+  .explore {
+    width: 78.54vw;
+    font-size: 5.18vw;
+    height: 85.97vw;
+  }
 }
