+++ frontend/src/index.scss
@@ -3,117 +3,163 @@ $card-spacing: 4.43vw;
 $text-color-gradient-1: #f9daa8;
 $text-color-gradient-2: #bc615e;
 $font-family: "Carta Marina W00 Regular";
+$medium-screen: 1023px;
+$small-screen: 76px;
 
 body {
-  margin: 0;
-  padding: 0;
+    margin: 0;
+    padding: 0;
 }
 
 .App {
-  display: flex;
-  flex-direction: column;
-  background-color: black;
-  padding-bottom: 7.84vw;
+    display: flex;
+    flex-direction: column;
+    background-color: black;
+    padding-bottom: 7.84vw;
 
-  .EventCards {
-    display: flex; 
-    flex-direction: row;
-    justify-content: center; 
-    flex-wrap: wrap; 
-    gap: 10px; 
-    padding-inline-start: 40px;
-    margin: auto;
-  }
+    .EventCards {
+        display: flex;
+        flex-direction: row;
+        justify-content: center;
+        flex-wrap: wrap;
+        gap: 10px;
+        margin: auto;
+
+        @media (max-width: $medium-screen) {
+            flex-direction: column;
+        }
+    }
 }
 
 .Events__heading {
-  font-size: xxx-large;
-  text-transform: uppercase;
-  background: linear-gradient($text-color-gradient-1, $text-color-gradient-2);
-  color: transparent;
-  -webkit-background-clip: text;
-  background-clip: text;
-  margin: auto;
-  font-family: $font-family;
+    font-size: xxx-large;
+    text-transform: uppercase;
+    background: linear-gradient($text-color-gradient-1, $text-color-gradient-2);
+    color: transparent;
+    -webkit-background-clip: text;
+    background-clip: text;
+    margin: auto;
+    font-family: $font-family;
 }
 
 // Single Card Component
 .EventCard {
-  display: flex; 
-  flex-direction: column;
-  width: auto;
-  margin-bottom: $card-margin;
+    display: flex;
+    flex-direction: column;
+    width: auto;
+    margin-bottom: $card-margin;
 
-  &.hiddenCard {
-    opacity: 0;
-    transform: translateY(90px) scale(0.9);
-    transition: all 0.5s;
-  }
-
-  &.showCard {
-    opacity: 1;
-    transform: translateY(0px) scale(1);
-  }
-
-  .eventContainer {
-    height: 100%;
-    padding: 1.5rem;
-    width: fit-content;
-    margin: 0 auto;
-
-    .event__img-container {
-      overflow: hidden;
-      width: 35.16vw;
-      height: 38.49vw;
-
-      .event__img {
-        margin: auto;
-        width: 100%;
-        height: 100%;
-        transition: all 0.8s;
-
-        &:hover {
-          transform: scale(1.1);
-        }
-      }
-    }
-
-    .event__content {
-      display: flex;
-      font-family: sans-serif;
-      flex-direction: column;
-      align-items: flex-start;
-      color: #fff;
-      width: 35.16vw;
-
-      .event__name {
-        margin-bottom: 0;
-        font-size: 2.5vw;
-        line-height: 1.51vw;
-        margin-top: 1.93vw;
-        font-weight: 50;
-        text-transform: uppercase;
-      }
-
-      .event__desc {
-        margin: 0;
-        font-size: 1.5vw;
-        color: rgb(144, 144, 144);
-        line-height: 0.99vw;
-        margin-top: 0.57vw;
-        font-weight: 100;
-      }
-    }
-  }
-
-  &:nth-child(2n) {
-    margin-left: $card-spacing;
     &.hiddenCard {
-        transform: translateY(calc(90px + 9.11vw)) scale(0.9);
+        opacity: 0;
+        transform: translateY(90px) scale(0.9);
+        transition: all 0.5s;
     }
-    
+
     &.showCard {
-        transform: translateY(9.11vw) scale(1);
+        opacity: 1;
+        transform: translateY(0px) scale(1);
     }
-  }
+
+    .event-container {
+        height: 100%;
+        padding: 1.5rem;
+        width: fit-content;
+        margin: 0 auto;
+
+        .event__img-container {
+            overflow: hidden;
+            width: 35.16vw;
+            height: 38.49vw;
+
+            @media (max-width: $medium-screen) {
+                width: 78.54vw;
+                height: 85.97vw;
+            }
+
+            .event__img {
+                margin: auto;
+                width: 100%;
+                height: 100%;
+                transition: all 0.8s;
+
+                &:hover {
+                    transform: scale(1.1);
+                }
+            }
+        }
+
+        .event__content {
+            display: flex;
+            font-family: sans-serif;
+            flex-direction: column;
+            align-items: flex-start;
+            color: #fff;
+            width: 35.16vw;
+
+            .event__name {
+                margin-bottom: 0;
+                font-size: 2.5vw;
+                line-height: 1.51vw;
+                margin-top: 1.93vw;
+                font-weight: 50;
+                text-transform: uppercase;
+
+                @media (max-width: $medium-screen) {
+                    font-size: 5.18vw;
+                    line-height: 3.37vw;
+                    margin-top: 4.31vw;
+                }
+            }
+
+            .event__desc {
+                margin: 0;
+                font-size: 1.5vw;
+                color: rgb(144, 144, 144);
+                line-height: 0.99vw;
+                margin-top: 0.57vw;
+                font-weight: 100;
+
+                @media (max-width: $medium-screen) {
+                    font-size: 3.09vw;
+                    line-height: 2.23vw;
+                    margin-top: 1.77vw;
+                }
+            }
+        }
+    }
+
+    @media (min-width: $medium-screen) {
+        &:nth-child(2n) {
+            margin-left: $card-spacing;
+
+            &.hiddenCard {
+                transform: translateY(calc(90px + 9.11vw)) scale(0.9);
+            }
+
+            &.showCard {
+                transform: translateY(9.11vw) scale(1);
+            }
+        }
+    }
+}
+
+.explore {
+    color: #fff;
+    width: 35.16vw;
+    height: 38.49vw;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-family: sans-serif;
+    font-size: 2.5vw;
+
+    &__text {
+        width: fit-content;
+    }
+
+    @media (max-width: $medium-screen) {
+        width: 78.54vw;
+        font-size: 5.18vw;
+        height: 85.97vw;
+    }
 }
